Core UI components to typescript and removing redux
