+++ app/App.tsx
@@ -1,22 +1,28 @@
-import { hot } from 'react-hot-loader/root';
-import React, { Component } from 'react';
+// import { hot } from 'react-hot-loader/root';
+import React, { useEffect, useState } from 'react';
+import { useHistory, Redirect, Route, Switch } from 'react-router-dom';
 import { Helmet } from 'react-helmet-async';
+import ReactGA from 'react-ga';
 import Intercom from 'react-intercom';
-import { connect } from 'react-redux';
-import Banner from './components/ui/Banner/Banner';
-import Navigation from './components/ui/Navigation/Navigation';
-// import CategoryPage from './find/FindPage';
-// import ResourcesTable from './search/ResourcesTable';
-import { round } from './utils/index';
-import whiteLabel from './utils/whitelabel';
-import 'react-select/dist/react-select.css';
+import { getLocation, COORDS_MID_SAN_FRANCISCO, AppContext, whiteLabel } from './utils';
+import { Banner, PopUpMessage, PopupMessageProp, HamburgerMenu, Navigation, UserWay } from './components/ui';
 import config from './config';
-import HamburgerMenu from './components/ui/HamburgerMenu';
-import { PopUpMessage } from './components/ui/PopUpMessage';
-import UserWay from './components/ui/UserWay';
-import { User } from './models';
-import Routes from './routes';
 import MetaImage from './assets/img/sfsg-preview.png';
+
+import { HomePage } from './pages/HomePage';
+import { AboutPage } from './pages/AboutPage';
+import { CovidPage } from './pages/CovidPage';
+import { PrivacyPolicyPage } from './pages/legal/PrivacyPolicy';
+import { TermsOfServicePage } from './pages/legal/TermsOfService';
+import { ListingDebugPage } from './pages/debug/ListingDemoPage';
+import { ResourceGuides, ResourceGuide } from './pages/ResourceGuides';
+import { RedirectToOrganizations, RedirectToOrganizationsEdit } from './pages/LegacyRedirects'
+import OrganizationEditPage from './pages/OrganizationEditPage';
+import { OrganizationListingPage } from './pages/OrganizationListingPage';
+import { SearchResultsPage } from './pages/SearchPage';
+import { ServiceListingPage } from './pages/ServiceListingPage';
+import ServiceDiscoveryForm from './pages/ServiceDiscoveryForm';
+import ServiceDiscoveryResults from './pages/ServiceDiscoveryResults';
 
 const {
   intercom,
@@ -26,126 +32,35 @@ const {
   title,
   userWay,
 } = whiteLabel;
+const outerContainerId = 'outer-container';
+const pageWrapId = 'page-wrap';
 
-const coordsInSanFrancisco = (coords: any) => {
-  // These are conservative bounds, extending into the ocean, the Bay, and Daly
-  // City.
-  const bb = {
-    top: 37.820633,
-    left: -122.562447,
-    bottom: 37.688167,
-    right: -122.326927,
-  };
-  return coords.lat > bb.bottom
-    && coords.lat < bb.top
-    && coords.lng > bb.left
-    && coords.lng < bb.right;
-};
+export const App = () => {
+  const history = useHistory()
+  const [hamburgerOpen, setHamburgerOpen] = useState(false);
+  const [popUpMessage, setPopUpMessage] = useState<PopupMessageProp>({ message: '', visible: false, type: 'success' });
+  const [userLocation, setUserLocation] = useState<any>(null);
 
-/**
- * Get location via HTML5 Geolocation API.
- */
-const getLocationBrowser = () => new Promise((resolve, reject) => {
-  if (navigator.geolocation) {
-    navigator.geolocation.getCurrentPosition(position => {
-      const coords = {
-        lat: round(position.coords.latitude, 4),
-        lng: round(position.coords.longitude, 4),
-      };
-      if (coordsInSanFrancisco(coords)) {
-        resolve(coords);
-      } else {
-        const msg = `User location out of bounds: ${coords.lat},${coords.lng}`;
-        console.log(msg);
-        reject(msg);
-      }
-    }, error => {
-      console.log(error);
-      reject(error);
-    });
-  } else {
-    const msg = 'Geolocation is not supported by this browser.';
-    console.log(msg);
-    reject(msg);
-  }
-});
+  useEffect(() => {
+    getLocation()
+      .then(loc => setUserLocation(loc))
+      .catch(err => {
+        console.log('Could not obtain location, defaulting to San Francisco.', err); // eslint-disable-line no-console
+        setUserLocation(COORDS_MID_SAN_FRANCISCO);
+      });
 
-/**
- * Get location via the Google Maps Geolocation API.
- */
-const getLocationGoogle = () => new Promise((resolve, reject) => {
-  // Results are not very accurate
-  let url = 'https://www.googleapis.com/geolocation/v1/geolocate';
-  if (config.GOOGLE_API_KEY) {
-    url += `?key=${config.GOOGLE_API_KEY}`;
-  }
-  fetch(url, { method: 'post' }).then(r => r.json())
-    .then(data => {
-      if (coordsInSanFrancisco(data.location)) {
-        resolve(data.location);
-      } else {
-        const msg = 'User location out of bounds';
-        console.log(msg);
-        reject(msg);
-      }
+    ReactGA.initialize(config.GOOGLE_ANALYTICS_ID);
+    return history.listen((loc) => {
+      const page = loc.pathname + loc.search;
+      ReactGA.set({ page });
+      ReactGA.pageview(loc.pathname);
+      window.scrollTo(0, 0);
     })
-    .catch(reject);
-});
+  }, []);
 
-/**
- * Get user location.
- *
- * Makes use of both the HTML5 Geolocation API and the Google Maps Geolocation
- * API. Currently restricts the location to within San Francisco to avoid
- * inaccurate geolocation results, but this should be removed if more locations
- * are added.
- *
- * @returns A Promise of a location, which is either an object with `lat` and
- * `lng` properties or an error if location is unavaible or out of bounds.
- */
-const getLocation = () => getLocationBrowser()
-  .catch(() => getLocationGoogle());
-
-class App extends Component {
-  state = {
-    hamburgerMenuIsOpen: false,
-  }
-
-  constructor(public props: any) {
-    super(props);
-    this.toggleHamburgerMenu = this.toggleHamburgerMenu.bind(this);
-    this.onHamburgerMenuStateChange = this.onHamburgerMenuStateChange.bind(this);
-  }
-
-  componentDidMount() {
-    const { setUserLocation } = this.props;
-    getLocation().then((coords: any) => {
-      setUserLocation(coords);
-    }).catch((e: any) => {
-      console.log('Could not obtain location, defaulting to San Francisco.', e);
-      // HACK: Hardcode middle of San Francisco
-      const userLocation = { lat: 37.7749, lng: -122.4194 };
-      setUserLocation(userLocation);
-    });
-  }
-
-  onHamburgerMenuStateChange(state: any) {
-    this.setState({ hamburgerMenuIsOpen: state.isOpen });
-  }
-
-  toggleHamburgerMenu() {
-    this.setState((state: any) => ({ hamburgerMenuIsOpen: !state.hamburgerMenuIsOpen }));
-  }
-
-  render() {
-    const { hamburgerMenuIsOpen } = this.state;
-    const { popUpMessage } = this.props;
-
-    const outerContainerId = 'outer-container';
-    const pageWrapId = 'page-wrap';
-
-    return (
-      <div id={outerContainerId}>
+  return (
+    <div id={outerContainerId}>
+      <AppContext.Provider value={{ setPopUpMessage, userLocation }}>
         <Helmet>
           <title>{title}</title>
           <meta property="og:url" content={siteUrl} />
@@ -163,39 +78,47 @@ class App extends Component {
         {userWay && <UserWay appID={config.SFFAMILIES_USERWAY_APP_ID} />}
         {intercom && config.INTERCOM_APP_ID && <Intercom appID={config.INTERCOM_APP_ID} />}
         <HamburgerMenu
-          isOpen={hamburgerMenuIsOpen}
+          isOpen={hamburgerOpen}
           outerContainerId={outerContainerId}
-          onStateChange={this.onHamburgerMenuStateChange}
+          onStateChange={(s: any) => setHamburgerOpen(s)}
           pageWrapId={pageWrapId}
-          toggleHamburgerMenu={this.toggleHamburgerMenu}
+          toggleHamburgerMenu={() => setHamburgerOpen(!hamburgerOpen)}
         />
         <div id={pageWrapId}>
           <Navigation
             showSearch={showSearch}
-            toggleHamburgerMenu={this.toggleHamburgerMenu}
+            toggleHamburgerMenu={() => setHamburgerOpen(!hamburgerOpen)}
           />
           {showBanner && <Banner />}
           <div className="container">
-            <Routes />
+            <Switch>
+              <Route exact path="/" component={HomePage} />
+              <Route exact path="/about" component={AboutPage} />
+              <Route exact path="/covid" component={CovidPage} />
+              <Route exact path="/demo/listing" component={ListingDebugPage} />
+              <Route exact path="/organizations/:id" component={OrganizationListingPage} />
+              <Route exact path="/organizations/:id/edit" component={OrganizationEditPage} />
+              <Route exact path="/organizations/new" component={OrganizationEditPage} />
+              <Route exact path="/privacy-policy" component={PrivacyPolicyPage} />
+              <Route exact path="/resource-guides" component={ResourceGuides} />
+              <Route exact path="/resource-guides/:id" component={ResourceGuide} />
+              <Route exact path="/search" component={SearchResultsPage} />
+              <Route exact path="/services/:id" component={ServiceListingPage} />
+              <Route exact path="/terms-of-service" component={TermsOfServicePage} />
+              <Route exact path="/:categorySlug/form" component={ServiceDiscoveryForm} />
+              <Route exact path="/:categorySlug/results" component={ServiceDiscoveryResults} />
+
+              {/* Legacy redirects */}
+              <Redirect path="/resource/new" to="/organizations/new" />
+              <Route exact path="/resource/edit" component={RedirectToOrganizationsEdit} />
+              <Route exact path="/resource" component={RedirectToOrganizations} />
+
+              <Redirect to="/" />
+            </Switch>
           </div>
           { popUpMessage && <PopUpMessage popUpMessage={popUpMessage} /> }
         </div>
-      </div>
-    );
-  }
-}
-
-function mapStateToProps(state: any) {
-  return {
-    userLocation: state.user.location,
-    popUpMessage: state.popUpMessage,
-  };
-}
-
-function mapDispatchToProps(dispatch: any) {
-  return {
-    setUserLocation: (location: any) => dispatch(User.setUserLocation(location)),
-  };
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(hot(App));
+      </AppContext.Provider>
+    </div>
+  );
+};
