+++ app/components/edit/MultiSelectDropdown.jsx
@@ -2,6 +2,7 @@ import React, { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
 import Select from 'react-select';
 import * as dataService from '../../utils/DataService';
+import 'react-select/dist/react-select.css';
 
 function dataToSelectValue(data) {
   return {
