+++ app/components/ui/Banner.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/utils/_helpers.scss';
+@import '../../styles/utils/_helpers.scss';
 
 .bannerContainer {
   width: 100%;
