--- app/components/ui/ScrollToTop.jsx
@@ -1,20 +1,0 @@
-// Copied from
-// https://github.com/ReactTraining/react-router/blob/v5.0.1/packages/react-router-dom/docs/guides/scroll-restoration.md
-import { Component } from 'react';
-import { withRouter } from 'react-router-dom';
-
-class ScrollToTop extends Component {
-  componentDidUpdate(prevProps) {
-    const { location } = this.props;
-    if (location !== prevProps.location) {
-      window.scrollTo(0, 0);
-    }
-  }
-
-  render() {
-    const { children } = this.props;
-    return children;
-  }
-}
-
-export default withRouter(ScrollToTop);
