--- app/components/ui/index.js
@@ -1,2 +1,0 @@
-export { default as Datatable } from './Datatable';
-export { default as Loader } from './Loader';
