--- app/config.js
@@ -1,22 +1,0 @@
-
-/* global CONFIG */
-const config = {
-  ALGOLIA_APPLICATION_ID: CONFIG.ALGOLIA_APPLICATION_ID,
-  ALGOLIA_INDEX_PREFIX: CONFIG.ALGOLIA_INDEX_PREFIX,
-  ALGOLIA_READ_ONLY_API_KEY: CONFIG.ALGOLIA_READ_ONLY_API_KEY,
-  GOOGLE_API_KEY: CONFIG.GOOGLE_API_KEY,
-  MOHCD_SUBDOMAIN: CONFIG.MOHCD_SUBDOMAIN,
-  MOHCD_DOMAIN: CONFIG.MOHCD_DOMAIN,
-  SFFAMILIES_DOMAIN: CONFIG.SFFAMILIES_DOMAIN,
-  LINKSF_DOMAIN: CONFIG.LINKSF_DOMAIN,
-  SFFAMILIES_USERWAY_APP_ID: '4EhNmEIwPz',
-  SENTRY_PROJECT_ID: '1291512',
-  SENTRY_PUBLIC_KEY: '64ab4b168902441a9e8069bab36849be',
-  INTERCOM_APP_ID: 'w50oz3tb',
-};
-
-if (CONFIG.TESTCAFE_RUNNING) {
-  delete config.INTERCOM_APP_ID;
-}
-
-export default config;
