--- app/store/configureStore.js
@@ -1,6 +1,0 @@
-/* eslint-disable */
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./configureStore.prod');
-} else {
-  module.exports = require('./configureStore.dev');
-}
