--- app/store/configureStore.prod.js
@@ -1,15 +1,0 @@
-import { createStore, applyMiddleware } from 'redux';
-import { routerMiddleware } from 'connected-react-router';
-import thunkMiddleware from 'redux-thunk';
-import { createBrowserHistory } from 'history';
-import createRootReducer from '../reducers';
-
-export const history = createBrowserHistory();
-
-export default function configureStore(initialState) {
-  return createStore(
-    createRootReducer(history),
-    initialState,
-    applyMiddleware(routerMiddleware(history), thunkMiddleware),
-  );
-}
