+++ app/utils/whitelabel.ts
@@ -6,7 +6,7 @@ import SFServiceLogo from '../assets/img/sf-service.svg';
 import SFSeal from '../assets/img/sf-seal.png';
 import LinkSFLogo from '../assets/img/link-sf.png';
 import config from '../config';
-import styles from '../components/ui/Navigation/Navigation.module.scss';
+import styles from '../components/ui/Navigation.scss';
 
 // Include new white label here
 type WhiteLabelSiteKey = 'defaultWhiteLabel' | 'SFServiceGuide' | 'SFFamilies' | 'LinkSF';
