+++ package.json
@@ -11,6 +11,7 @@
     "@babel/polyfill": "^7.4.4",
     "@sentry/browser": "^4.0.6",
     "@typeform/embed": "^0.13.1",
+    "@types/react-burger-menu": "^2.8.1",
     "classnames": "^2.2.6",
     "connected-react-router": "^6.5.2",
     "copy-webpack-plugin": "^5.1.1",
