+++ tsconfig.json
@@ -14,9 +14,6 @@
     "esModuleInterop": true, // Allows us to use default imports with CommonJS modules (which some third-party packages use)
     "skipLibCheck": true // Don't type check .d.ts files, which mostly come from third party libraries.
   },
-  "paths": {
-    "*": ["app/index.d.ts"]
-  },
   "include": ["app", "cypress"],
   "exclude": ["node_modules"]
 }
